packages/agent-kit/beginner/bitcoin-query/pyproject.toml
[project]
name = "unisat-bitcoin-query-agent"
version = "0.1.0"
description = "UniSat Bitcoin Query Agent - A simple blockchain query agent"
authors = [
    {name = "UniSat", email = "dev@unisat.io"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "agentkit>=0.1.0",
    "veadk>=0.1.0",
    "google-adk>=0.1.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
